requirements.txt
requests
python-dotenv
lxml
pandas
pytrends